+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Misc:
 - **Flake8** Add `parallel` option [#2130](https://github.com/sider/runners/pull/2130)
 - **detekt** Add `parallel` and `target` options [#2131](https://github.com/sider/runners/pull/2131)
 - **PHP_CodeSniffer** Re-consider options (`target`, `standard`, `parallel`) [#2132](https://github.com/sider/runners/pull/2132)
+- **Checkstyle** Enable new recommended configuration [#2137](https://github.com/sider/runners/pull/2137)
 - **Metrics Code Clone** Report the total number of duplicated lines [#2146](https://github.com/sider/runners/pull/2146)
 - Install default npm dependencies via npm 7 [#2150](https://github.com/sider/runners/pull/2150)
 - **Clang-Tidy** Loosen the version constraint for clang-tidy [#2153](https://github.com/sider/runners/pull/2153)